requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
openai==0.28.1
dspy-ai==2.1.4 